vm::object::payload
