+++ vm/src/object/mod.rs
@@ -1,5 +1,7 @@
 mod core;
 mod ext;
+mod payload;
 
 pub use self::core::*;
 pub use self::ext::*;
+pub use self::payload::*;
